ensure msg in infobarStopper is right type
When embedded in other extensions, the infobarStopper request handler was responding to messages sent from outside extension code.
This caused a bug in Gombot where CompleteInfobar would have to be called twice to dismiss an infobar, meaning the user would have to enter the last digit of their PIN a second time.